YoutubeChat/requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
youtube-transcript-api
faiss-cpu
sentence-transformers
python-dotenv